requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
pyparsing==2.4.7
torchvision==0.9.1